handle an empty slot in the signature array
